+++ src/main/java/net/spell_power/mixin/LivingEntityMixin.java
@@ -4,14 +4,17 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.DefaultAttributeContainer;
+import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.world.World;
 import net.spell_power.SpellPowerMod;
 import net.spell_power.api.SpellPowerMechanics;
+import net.spell_power.api.SpellResistance;
 import net.spell_power.api.SpellSchools;
 import net.spell_power.config.AttributesConfig;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LivingEntity.class)
@@ -36,6 +39,18 @@ abstract class LivingEntityMixin extends Entity {
                     info.getReturnValue().add(attribute);
                 }
             }
+            for (var resistance: SpellResistance.Attributes.all) {
+                info.getReturnValue().add(resistance.attribute);
+            }
         }
+    }
+
+    @ModifyVariable(method = "damage", at = @At("HEAD"), ordinal = 0)
+    private float damage_resistance(float amount, DamageSource source) {
+        var entity = (LivingEntity)(Object)this;
+        if (entity.isInvulnerableTo(source) || entity.isDead()) {
+            return amount;
+        }
+        return (float) SpellResistance.resist(entity, amount, source);
     }
 }